clean up; minor fixes
