+++ ckg/report_manager/pages/adminPage.py
@@ -9,12 +9,27 @@ description = ""
 dash.register_page(__name__, path='/apps/admin', title=f"{title} - {subtitle}", description=description)
 
 
-def layout():
-    # TODO: adapt authentication for admin
+def layout(new_user=None, error_new_user=None, running=None):
     session_cookie = flask.request.cookies.get('custom-auth-session')
     logged_in = session_cookie is not None
     if logged_in == False:
         return html.Div(["Please ", dcc.Link("login", href="/apps/loginPage"), " to continue"])
+
+    action_status_layout = []
+    if new_user != None:
+        if error_new_user != None:
+            action_status_layout.append(
+                html.Div(children=[html.H3("– Error creating new user: {} – ".format(new_user.replace('%20', ' ')))],
+                         className='error_panel'))
+        else:
+            action_status_layout.append(
+                html.Div(children=[html.H3("– New user successfully created: {} –".format(new_user))],
+                         className='info_panel'))
+
+    if running != None:
+        action_status_layout.append(html.Div(children=[html.H3(
+            "– The {} update is running. This will take a while, check the logs: graphdb_builder.log for more information –".format(
+                running))], className='info_panel'))
 
     create_user_form = [html.H3("Create CKG User"), html.Form([
         html.Div(children=[html.Label('Name'),
@@ -45,16 +60,12 @@ def layout():
                            html.Div(children=[
                                html.Form([html.Button('Full Update', type='submit', className='button_link'),
                                           html.Div(children=[html.H4("Download:"),
-                                                             html.Label("Yes", className='radioitem'),
-                                                             dcc.Input(id='Yes',
-                                                                       name='dwn-radio',
-                                                                       value=True,
-                                                                       type='radio'),
-                                                             html.Label("No", className='radioitem'),
-                                                             dcc.Input(id='No',
-                                                                       name='dwn-radio',
-                                                                       value=False,
-                                                                       type='radio')])], action='/update_full',
+                                                             dcc.RadioItems(options=[
+                                                                 {'label': 'Yes', 'value': True},
+                                                                 {'label': 'No', 'value': False}],
+                                                                 value=False, inline=True, className='dwn-radio'),
+                                                             ])],
+                                         action='/update_full',
                                          method='post'),
                                html.P(
                                    "This option will regenerate the entire database, downloading data from the different Ontologies and Databases (Download=Yes) and loading them and all existing projects into CKG's graph database.",
@@ -68,5 +79,6 @@ def layout():
         html.Div(children=description),
     ])]
 
-    admin_layout.extend(admin_options)
+    admin_layout.append(admin_options)
+    admin_layout.append(html.Div(children=action_status_layout))
     return admin_layout
